aspeed: Zero extend flash files to 128MB
This finally removes the requirement that you have to pad your QEMU
flash images with zeroes.

Now QEMU will automatically zero-extend your flash device file to the
Aspeed FMC's segment size if it is smaller than it should be.

For example:

    $ ls -lh flash-yosemite35
    -rw-r--r-- 2 pdel users 30M Jun 17 12:23 flash-yosemite35
    $ qemu-system-arm -machine fby35-bmc -drive file=flash-yosemite35,format=raw,if=mtd -nographic
    $ ls -lh flash-yosemite35
    -rw-r--r-- 2 pdel users 128M Jun 17 12:25 flash-yosemite35

If your file is large enough already, then QEMU will not truncate it.

For example, if your image is 256MB, it won't truncate to 128MB.

    $ dd if=/dev/zero of=tmp.mtd bs=1M count=256
    256+0 records in
    256+0 records out
    268435456 bytes (268 MB, 256 MiB) copied, 0.125099 s, 2.1 GB/s
    pdel@devvm9194:~/local/qemu (blk)
    $ ./build/qemu-system-arm -machine fby35-bmc -drive file=tmp.mtd,format=raw,if=mtd -nographic
    QEMU 7.0.50 monitor - type 'help' for more information
    (qemu) q
    pdel@devvm9194:~/local/qemu (blk)
    $ ls -lh tmp.mtd
    -rw-r--r-- 1 pdel users 256M Jun 17 15:46 tmp.mtd

Signed-off-by: Peter Delevoryas <pdel@fb.com>