README: Add instructions for booting npcm845-evb
Also updated the bios with an image I built. This
will only work with the test image pointed to in
the README:

wget https://github.com/peterdelevoryas/openbmc/releases/download/nv1/npcm845.mtd

Also, this panics in psci_dt_init right now, probably because we're
using the Cortex-A53 instead of the Cortex-A35. But that should be fixed
soon.